+++ voxbrix_common/src/entity/action.rs
@@ -5,7 +5,7 @@ use serde::{
 };
 
 #[derive(Serialize, Deserialize, PartialEq, Eq, PartialOrd, Hash, Ord, Copy, Clone, Debug)]
-pub struct Action(pub u64);
+pub struct Action(pub u32);
 
 impl nohash_hasher::IsEnabled for Action {}
